remarks applied, tailwindcss must be resolved
