+++ frontend/src/pages/about/About.tsx
@@ -1,4 +1,4 @@
-import { Button, Divider, SvgIconTypeMap, SxProps, Typography } from '@mui/material';
+import { Alert, Button, Link, Divider, SvgIconTypeMap, SxProps, Typography } from '@mui/material';
 import PageHeader from '../../components/PageHeader';
 import PageSection from '../../components/PageSection';
 import Logo from '../../components/logo/Logo';
@@ -19,8 +19,8 @@ import CardfightVanguardImage from '../../../assets/games/cardfightvanguard.png'
 import React, { HTMLAttributeAnchorTarget, useState } from 'react';
 import { OverridableComponent } from '@mui/material/OverridableComponent';
 import SupportedGamesImage from '../../components/about/SupportedGamesImage';
-import { useTranslation } from 'react-i18next';
-import BetaAlertDiv from '../../components/about/BetaAlertDiv';
+import { Trans, useTranslation } from 'react-i18next';
+import CloseIcon from '@mui/icons-material/Close';
 
 type CardflowReasons = {
   logoColor: string;
@@ -46,11 +46,49 @@ function About(): JSX.Element {
     },
   ];
 
+  const handleClose = () => {
+    setAlertVisible(false);
+  };
+
   return (
     <section className="bg-[#F5F5F5] min-h-[100vh] pb-4">
       <CardflowTabs />
       <PageHeader heading={t('main.title')} />{' '}
-      {alertVisible && <BetaAlertDiv setAlertVisible={setAlertVisible} />}
+      {alertVisible && (
+        <Alert
+          id="closable-alert"
+          severity="info"
+          action={
+            <Button
+              aria-label="close"
+              sx={{ borderRadius: '8px', minWidth: '24px' }}
+              onClick={handleClose}
+            >
+              <CloseIcon sx={{ fontSize: '16px' }} />
+            </Button>
+          }
+          sx={{
+            fontSize: '18px',
+            width: '83.33%',
+            margin: '16px auto',
+            bgcolor: '#E5F6FD',
+            display: 'flex',
+            color: '#000',
+            fontWeight: '300',
+            border: '1px solid #666',
+            borderRadius: '8px',
+            alignItems: 'center',
+            '.MuiAlert-icon': {
+              fontSize: 30,
+            },
+          }}
+        >
+          <Trans i18nKey="betaWarning">
+            {t('main.betaWarning')}
+            <Link href="/about/contact" />
+          </Trans>
+        </Alert>
+      )}
       <div className="w-5/6 mx-auto my-4">
         <PageSection className="p-8 my-4">
           <h2 className="text-2xl mb-12 font-bold text-center lg:text-left">
