+++ frontend/src/translations/bg/about.json
@@ -5,10 +5,7 @@
   },
   "main": {
     "title": "За нас",
-    "betaWarning": "Приложението все още е в бета версия! Ако откриете грешки, не се колебайте да ги докладвате.",
-    "betaWarningFirstPart": "Приложението все още е в бета версия! Ако откриете грешки, не се колебайте да ги",
-    "betaWarningSecondPart": "",
-    "report": "докладвате.",
+    "betaWarning": "Приложението все още е в бета версия! Ако откриете грешки, не се колебайте да ги <1>докладвате.</1>",
     "whyCardflow": "Защо да избереш Cardflow?",
     "reasons": {
       "free": "Безплатен",
