+++ frontend/src/translations/en/about.json
@@ -5,9 +5,7 @@
   },
   "main": {
     "title": "About",
-    "betaWarningFirstPart": "The app is still in beta! If you find any bugs, don't hesitate to",
-    "betaWarningSecondPart": "them.",
-    "report": "report",
+    "betaWarning": "The app is still in beta! If you find any bugs, don't hesitate to <1>report</1> them.",
     "whyCardflow": "Why Cardflow?",
     "reasons": {
       "free": "Free",
